+++ tests/test_handlers.py
@@ -4,18 +4,23 @@
 import json
 import os
 import re
+from pathlib import Path
+from tempfile import TemporaryDirectory
 from unittest import mock
 
 from eodag import SearchResult
+from eodag import __version__ as eodag_version
 from eodag.api.core import DEFAULT_ITEMS_PER_PAGE
 from eodag.types.queryables import QueryablesDict
 from notebook.notebookapp import NotebookApp
 from shapely.geometry import shape
-from tornado.testing import AsyncHTTPTestCase
+from tornado.httpclient import HTTPClientError
+from tornado.testing import AsyncHTTPTestCase, gen_test
 from tornado.web import authenticated
 
+from eodag_labextension import __version__ as labextension_version
 from eodag_labextension import load_jupyter_server_extension
-from eodag_labextension.handlers import APIHandler
+from eodag_labextension.handlers import APIHandler, get_eodag_api, set_conf_symlink
 
 
 class MockUser:
@@ -40,6 +45,21 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
                 os.environ.pop(k)
         os.environ.update(cls.eodag_env_backup)
 
+    def setUp(self):
+        super().setUp()
+        self.patcher_xsrf = mock.patch.object(APIHandler, "check_xsrf_cookie", return_value=MockUser())
+        self.patcher_user = mock.patch.object(APIHandler, "get_current_user", return_value=MockUser())
+        self.patcher_auth = mock.patch.object(authenticated, "__call__", return_value=lambda x: x)
+        self.mock_xsrf = self.patcher_xsrf.start()
+        self.mock_user = self.patcher_user.start()
+        self.mock_auth = self.patcher_auth.start()
+
+    def tearDown(self):
+        super().tearDown()
+        self.patcher_xsrf.stop()
+        self.patcher_user.stop()
+        self.patcher_auth.stop()
+
     def get_app(self):
         # Create a new NotebookApp instance
         app = NotebookApp()
@@ -50,119 +70,121 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
 
         return app.web_app
 
-    @mock.patch.object(APIHandler, "check_xsrf_cookie", return_value=MockUser())
-    @mock.patch.object(APIHandler, "get_current_user", return_value=MockUser())
-    @mock.patch.object(authenticated, "__call__", return_value=lambda x: x)
-    def fetch_results(self, url, mock_auth, mock_user, mock_xsrf, **kwargs):
+    async def fetch_results(self, url, **kwargs):
         """Check that request status is 200 and return the json result as dict"""
-        response = self.fetch(url, **kwargs)
+        response = await self.http_client.fetch(self.get_url(url), **kwargs)
         self.assertEqual(response.code, 200)
         return json.loads(response.body.decode("utf-8"))
 
-    @mock.patch.object(APIHandler, "check_xsrf_cookie", return_value=MockUser())
-    @mock.patch.object(APIHandler, "get_current_user", return_value=MockUser())
-    @mock.patch.object(authenticated, "__call__", return_value=lambda x: x)
-    def fetch_results_error(
-        self,
-        url,
-        error_code=None,
-        mock_auth=None,
-        mock_user=None,
-        mock_xsrf=None,
-        **kwargs,
-    ):
+    async def fetch_results_error(self, url, error_code=None, **kwargs):
         """Check that request returns a 400 error"""
-        response = self.fetch(url, **kwargs)
-        self.assertNotEqual(response.code, 200)
-        if error_code:
-            self.assertEqual(response.code, error_code)
-        return json.loads(response.body.decode("utf-8"))
+        # with self.assertRaises(HTTPClientError) as err:
+        try:
+            response = await self.http_client.fetch(self.get_url(url), **kwargs)
+        except HTTPClientError as err:
+            if error_code:
+                self.assertEqual(err.code, error_code)
+        else:
+            self.fail(f"Expected {response} to raise HTTP {error_code}")
 
-    def test_product_types(self):
+    @gen_test
+    async def test_product_types(self):
         # all product types
-        results = self.fetch_results("/eodag/product-types")
+        results = await self.fetch_results("/eodag/product-types")
         self.assertIn("S2_MSI_L1C", [pt["ID"] for pt in results])
 
         # single provider product types
-        less_results = self.fetch_results("/eodag/product-types?provider=peps")
+        less_results = await self.fetch_results("/eodag/product-types?provider=peps")
         self.assertGreater(len(less_results), 0)
         self.assertLess(len(less_results), len(results))
 
         # unknown provider
-        self.fetch_results_error("/eodag/product-types?provider=foo", 400)
+        await self.fetch_results_error("/eodag/product-types?provider=foo", 400)
 
-    def test_providers(self):
+    @gen_test
+    async def test_providers(self):
         # all providers
-        results = self.fetch_results("/eodag/providers")
+        results = await self.fetch_results("/eodag/providers")
         self.assertIn("peps", [res["provider"] for res in results])
 
-        less_results = self.fetch_results("/eodag/providers?product_type=S2_MSI_L1C")
+        less_results = await self.fetch_results("/eodag/providers?product_type=S2_MSI_L1C")
         self.assertGreater(len(less_results), 0)
         self.assertLess(len(less_results), len(results))
 
-        result_with_name = self.fetch_results("/eodag/providers?keywords=peps")
+        result_with_name = await self.fetch_results("/eodag/providers?keywords=peps")
         self.assertEqual(len(result_with_name), 1)
         self.assertEqual(result_with_name[0]["provider"], "peps")
 
-        result_with_description = self.fetch_results("/eodag/providers?keywords=cop")
+        result_with_description = await self.fetch_results("/eodag/providers?keywords=cop")
         self.assertGreater(len(result_with_description), 2)
         providers = [r["provider"] for r in result_with_description]
         self.assertIn("cop_marine", providers)
 
-        no_result = self.fetch_results("/eodag/providers?product_type=foo")
+        no_result = await self.fetch_results("/eodag/providers?product_type=foo")
         self.assertEqual(len(no_result), 0)
 
-    def test_guess_product_types(self):
-        all_results = self.fetch_results("/eodag/guess-product-type")
+    @gen_test
+    async def test_guess_product_types(self):
+        all_results = await self.fetch_results("/eodag/guess-product-type")
         self.assertIn("S2_MSI_L1C", [pt["ID"] for pt in all_results])
+        self.assertListEqual(sorted(list(all_results[0].keys())), ["ID", "title"])
 
-        one_provider_results = self.fetch_results("/eodag/guess-product-type?provider=creodias")
+        one_provider_results = await self.fetch_results("/eodag/guess-product-type?provider=creodias")
         self.assertLess(len(one_provider_results), len(all_results))
         self.assertIn("COP_DEM_GLO90_DGED", [pt["ID"] for pt in all_results])
+        self.assertListEqual(sorted(list(one_provider_results[0].keys())), ["ID", "title"])
 
-        one_result = self.fetch_results("/eodag/guess-product-type?keywords=S2_MSI_L1C")
+        one_result = await self.fetch_results("/eodag/guess-product-type?keywords=S2_MSI_L1C")
         self.assertEqual(len(one_result), 1)
         self.assertEqual(one_result[0]["ID"], "S2_MSI_L1C")
+        self.assertListEqual(sorted(list(one_result[0].keys())), ["ID", "title"])
 
-        another_result = self.fetch_results("/eodag/guess-product-type?keywords=Sentinel2%20L1C")
+        another_result = await self.fetch_results("/eodag/guess-product-type?keywords=Sentinel2%20L1C")
         self.assertEqual(len(another_result), 1)
         self.assertEqual(another_result[0]["ID"], "S2_MSI_L1C")
+        self.assertListEqual(sorted(list(another_result[0].keys())), ["ID", "title"])
 
-        more_results = self.fetch_results("/eodag/guess-product-type?keywords=Sentinel")
+        more_results = await self.fetch_results("/eodag/guess-product-type?keywords=Sentinel")
         self.assertGreater(len(more_results), 1)
         self.assertLess(len(more_results), len(all_results))
         self.assertIn("S2_MSI_L1C", [pt["ID"] for pt in more_results])
+        self.assertListEqual(sorted(list(more_results[0].keys())), ["ID", "title"])
 
-        less_results = self.fetch_results("/eodag/guess-product-type?keywords=Sentinel&provider=peps")
+        less_results = await self.fetch_results("/eodag/guess-product-type?keywords=Sentinel&provider=peps")
         self.assertGreater(len(more_results), 1)
         self.assertLess(len(less_results), len(more_results))
         self.assertEqual(less_results[0]["ID"], "S1_SAR_GRD")
+        self.assertListEqual(sorted(list(less_results[0].keys())), ["ID", "title"])
 
-        other_results = self.fetch_results("/eodag/guess-product-type?keywords=cop")
+        other_results = await self.fetch_results("/eodag/guess-product-type?keywords=cop")
         self.assertGreater(len(other_results), 1)
         self.assertLess(len(other_results), len(all_results))
         self.assertTrue(other_results[0]["ID"].lower().startswith("cop"))
+        self.assertListEqual(sorted(list(other_results[0].keys())), ["ID", "title"])
 
-        self.fetch_results_error("/eodag/guess-product-type?provider=foo", 400)
+        await self.fetch_results_error("/eodag/guess-product-type?provider=foo", 400)
 
-    def test_get_not_found(self):
-        self.fetch_results_error("/eodag/foo", 404)
+    @gen_test
+    async def test_get_not_found(self):
+        await self.fetch_results_error("/eodag/foo", 404)
 
-    def test_post_not_found(self):
-        self.fetch_results_error("/eodag/foo/bar", 404, method="POST", body=json.dumps({}))
+    @gen_test
+    async def test_post_not_found(self):
+        await self.fetch_results_error("/eodag/foo/bar", 404, method="POST", body=json.dumps({}))
 
     @mock.patch(
         "eodag.api.core.EODataAccessGateway.search",
         autospec=True,
         return_value=SearchResult([], 0),
     )
-    def test_search(self, mock_search):
+    @gen_test
+    async def test_search(self, mock_search):
         geom_dict = {
             "type": "Polygon",
             "coordinates": [[[0, 2], [0, 3], [1, 3], [1, 2], [0, 2]]],
         }
         # full example
-        result = self.fetch_results(
+        result = await self.fetch_results(
             "/eodag/S2_MSI_L1C",
             method="POST",
             body=json.dumps(
@@ -205,7 +227,7 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
 
         # minimal example
         mock_search.reset_mock()
-        result = self.fetch_results(
+        result = await self.fetch_results(
             "/eodag/S2_MSI_L1C",
             method="POST",
             body=json.dumps({}),
@@ -218,7 +240,7 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
 
         # date error
         mock_search.reset_mock()
-        self.fetch_results_error(
+        await self.fetch_results_error(
             "/eodag/S2_MSI_L1C",
             400,
             method="POST",
@@ -227,7 +249,7 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
 
         # geom error
         mock_search.reset_mock()
-        self.fetch_results_error(
+        await self.fetch_results_error(
             "/eodag/S2_MSI_L1C",
             400,
             method="POST",
@@ -239,8 +261,9 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
         autospec=True,
         return_value=QueryablesDict(additional_properties=False),
     )
-    def test_queryables(self, mock_list_queryables):
-        results = self.fetch_results(
+    @gen_test
+    async def test_queryables(self, mock_list_queryables):
+        results = await self.fetch_results(
             "/eodag/queryables?"
             "provider=some_provider&productType=some_product_type"
             "&param1=paramValue1&param2=paramValue2"
@@ -255,3 +278,37 @@ class TestEodagLabExtensionHandler(AsyncHTTPTestCase):
             param1="paramValue1",
             param2="paramValue2",
         )
+
+    @gen_test
+    async def test_info(self):
+        infos = await self.fetch_results("/eodag/info")
+        self.assertIn("packages", infos)
+        self.assertEqual(infos["packages"]["eodag"]["version"], eodag_version)
+        self.assertEqual(infos["packages"]["eodag_labextension"]["version"], labextension_version)
+
+    @mock.patch.dict(os.environ, {"EODAG_LABEXTENSION__DEBUG": "true"})
+    @gen_test
+    async def test_debug(self):
+        infos = await self.fetch_results("/eodag/info")
+        self.assertTrue(infos["debug"])
+
+    @gen_test(timeout=120)
+    async def test_set_conf_symlink(self):
+        with TemporaryDirectory() as tmpdir:
+            custom_cfg_file = Path(tmpdir) / "foo.yml"
+            custom_cfg_file.touch()
+            custom_cfg_file_str = str(custom_cfg_file.absolute())
+            local_cfg_file = Path("eodag-config") / "eodag.yml"
+
+            # custom config file
+            with mock.patch.dict(os.environ, {"EODAG_CFG_FILE": custom_cfg_file_str}):
+                eodag_api = await get_eodag_api()
+                set_conf_symlink(eodag_api)
+                self.assertTrue(os.path.isdir("eodag-config"))
+                self.assertTrue(os.path.islink(str(local_cfg_file)))
+                self.assertEqual(custom_cfg_file, Path(os.readlink(str(local_cfg_file))))
+
+            # default config file
+            set_conf_symlink(eodag_api)
+            self.assertTrue(os.path.islink("eodag-config"))
+            self.assertEqual(Path(eodag_api.conf_dir) / "eodag.yml", Path(os.readlink("eodag-config")) / "eodag.yml")